restructuring
- changed ClubService to an interface
- changed references of User in club to UserProfile
