+++ backend/src/main/java/com/crashcourse/kickoff/tms/club/repository/PlayerApplicationRepository.java
@@ -2,9 +2,9 @@ package com.crashcourse.kickoff.tms.club.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.crashcourse.kickoff.tms.club.model.PlayerApplication;
-import com.crashcourse.kickoff.tms.user.model.User;
+import com.crashcourse.kickoff.tms.user.model.PlayerProfile;
 import com.crashcourse.kickoff.tms.club.Club;
 
 public interface PlayerApplicationRepository extends JpaRepository<PlayerApplication, Long> {
-    boolean existsByPlayerAndClub(User player, Club club);  // Should accept User and Club
+    boolean existsByPlayerProfileAndClub(PlayerProfile playerProfile, Club club);  // Should accept User and Club
 }
