+++ backend/src/main/java/com/crashcourse/kickoff/tms/user/model/User.java
@@ -77,8 +77,6 @@ public class User implements UserDetails {
         this.username = username;
         this.password = password;
         this.roles = roles;
-        this.playerProfile = playerProfile;
-        this.hostProfile = hostProfile;
     }
 
     /*x
