+++ backend/src/main/java/com/crashcourse/kickoff/tms/user/service/PlayerProfileService.java
@@ -1,10 +1,11 @@
 package com.crashcourse.kickoff.tms.user.service;
 
 import java.util.*;
+
+import com.crashcourse.kickoff.tms.user.model.PlayerPosition;
 import com.crashcourse.kickoff.tms.user.model.PlayerProfile;
 
 public interface PlayerProfileService {
-    List<PlayerProfile> getPlayers();
-    PlayerProfile addPlayer(PlayerProfile player);
-    PlayerProfile getPlayer(Long id);  
+    List<PlayerProfile> getPlayerProfiles();
+    PlayerProfile updatePlayerPosition(Long userId, PlayerPosition preferredPosition);
 }
