--- backend/src/main/java/com/crashcourse/kickoff/tms/user/service/PlayerServiceImpl.java
@@ -1,5 +1,0 @@
-package com.crashcourse.kickoff.tms.user.service;
-
-public class PlayerServiceImpl {
-    
-}
