+++ backend/src/main/resources/static/ClubTest.http
@@ -27,5 +27,6 @@ Authorization: Basic admin:password
 Content-Type: application/json
 
 {
+  "playerProfileId": 1,
   "desiredPosition": "POSITION_FORWARD"
 }
